+++ sdk/schemaregistry/schema-registry/recordings/node/schemaregistryclient/recording_rejects_schema_registration_with_invalid_args.js
@@ -1,6 +1,6 @@
 let nock = require('nock');
 
-module.exports.hash = "590d8b8e9e17f48ff08c6808bb01e9b9";
+module.exports.hash = "e6c726bef917a02ff50524896e0464f2";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
@@ -23,19 +23,19 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
   'x-ms-request-id',
-  'f2dbdde5-db5a-4996-bc26-6619bda96302',
+  'fcc056d4-8fbe-4bac-aa2a-7cb5800a2600',
   'x-ms-ests-server',
-  '2.1.11898.12 - SCUS ProdSlices',
+  '2.1.12071.13 - EUS ProdSlices',
   'Set-Cookie',
-  'fpc=ArqRweA9seFErsvWrKI3q4k; expires=Fri, 03-Sep-2021 20:29:18 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=Ajc8F7qBb6xFgZ4OjjUppok; expires=Sun, 24-Oct-2021 16:46:21 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
-  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7WevrB39Hpn2A5cIQkGUHfSepYJHOBhvl3p6KmbflGZD6u0pjSPxQdV4kZGo6nFah8iHd-4k-84MigLo3GJinog6Xra9FhWceFc74WVKzAth_txfufIAzb9jY7vrwsEp5yt3MUYClJIxFR0IeNBFKEDK8ujhuBWwNjXePAV3U4LYVZl4gAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
+  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7WevrOdhqbbSAe5sToY8FG1hWoJbVz4fRMV76Yb6HuaaHQSz6mi2n64GqHkgblGlblC7Kb132ot0QYWEs8sRkehQ1Nz-oqz5doUy_RbwxC4aNRPwg_SxfNKcr10_ZZznpWdoYaGo54lxeNhMBjT_dbBiDIU3PDUEd6TR7mRqhcd6KOQMgAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
   'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
   'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
   'Date',
-  'Wed, 04 Aug 2021 20:29:18 GMT',
+  'Fri, 24 Sep 2021 16:46:21 GMT',
   'Content-Length',
   '980'
 ]);
@@ -58,25 +58,25 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
   'x-ms-request-id',
-  '288acde2-7e5e-495a-b5d8-0a1eb3414301',
+  '9f1cdf8d-591b-4115-8f39-413b3e3f2800',
   'x-ms-ests-server',
-  '2.1.11898.12 - NCUS ProdSlices',
+  '2.1.12071.13 - NCUS ProdSlices',
   'Set-Cookie',
-  'fpc=ArqRweA9seFErsvWrKI3q4k; expires=Fri, 03-Sep-2021 20:29:18 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=AhX_3H1yLtBPv32sOAswBww; expires=Sun, 24-Oct-2021 16:46:21 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
-  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7WevriBywxOST0AVd0hJaHrDK7wzFeSP88z5rferb3501SkvY9kA2tjUIsJ_IEzEq6qdbIo5dxdwMBW9TeGJ8pk-6bmFgGXz1s-d3vTHiW4LxQ9c6MXgSdsbXSgGcv7s5ttd_9IGyJhL4C1K3UqWab_TqwdRymEoOVjIKJWrIPY3B-IsgAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
+  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7WevrU_RO5a_fJ8xIMd9bP1pz5bUkH-Zu8AlOjZ7frXCDBp-6oCpi75urw7agfNEpTJLYpRgNWpsHPWMdKKSCePoxV4NKx2ghdgna4PgOIbQyW8JOVNoOk0eWhzsSWWoMPjFjiBNiHxXgfgTkN4lJjXeUSr9DpycB4tJcAki4Drm-REQgAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
   'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
   'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
   'Date',
-  'Wed, 04 Aug 2021 20:29:18 GMT',
+  'Fri, 24 Sep 2021 16:46:21 GMT',
   'Content-Length',
   '1753'
 ]);
 
 nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
-  .post('/azuretenantid/oauth2/v2.0/token', "client_id=azure_client_id&scope=https%3A%2F%2Fsanitized%2F&grant_type=client_credentials&x-client-SKU=msal.js.node&x-client-VER=1.2.0&x-client-OS=linux&x-client-CPU=x64&x-ms-lib-capability=retry-after, h429&x-client-current-telemetry=2|771,0|,&x-client-last-telemetry=2|0|||0,0&client-request-id=6ab86dfd-93c2-4d0e-8120-1badd766e446&client_secret=azure_client_secret")
+  .post('/azuretenantid/oauth2/v2.0/token', "client_id=azure_client_id&scope=https%3A%2F%2Fsanitized%2F&grant_type=client_credentials&x-client-SKU=msal.js.node&x-client-VER=1.3.1&x-client-OS=linux&x-client-CPU=x64&x-ms-lib-capability=retry-after, h429&x-client-current-telemetry=5|771,2,,,|,&x-client-last-telemetry=5|0|||0,0&client-request-id=f9c04f0b-c3e4-4d1f-b5c6-d33ae4ae5ad9&client_secret=azure_client_secret&claims=%7B%22access_token%22%3A%7B%22xms_cc%22%3A%7B%22values%22%3A%5B%22CP1%22%5D%7D%7D%7D")
   .reply(200, {"token_type":"Bearer","expires_in":86399,"ext_expires_in":86399,"access_token":"access_token"}, [
   'Cache-Control',
   'no-store, no-cache',
@@ -93,19 +93,19 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
   'x-ms-request-id',
-  '8089a432-0de7-4720-afe5-62162b610000',
+  '2200c9bd-6656-4be8-9c16-a373a11b3c00',
   'x-ms-ests-server',
-  '2.1.11935.12 - SCUS ProdSlices',
+  '2.1.12071.13 - SCUS ProdSlices',
   'x-ms-clitelem',
   '1,0,0,,',
   'Set-Cookie',
-  'fpc=ArqRweA9seFErsvWrKI3q4nmLYaiAQAAAJ7vnNgOAAAA; expires=Fri, 03-Sep-2021 20:29:18 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=AoO4Y7wEgcVDmjL2AUyvzX7mLYaiAQAAAN3339gOAAAA; expires=Sun, 24-Oct-2021 16:46:21 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
   'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
   'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
   'Date',
-  'Wed, 04 Aug 2021 20:29:18 GMT',
+  'Fri, 24 Sep 2021 16:46:21 GMT',
   'Content-Length',
   '1321'
 ]);
@@ -113,7 +113,7 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
 nock('https://endpoint', {"encodedQueryParams":true})
   .put('/$schemagroups/group-1/schemas/azsdk_js_test', {"type":"record","name":"User","namespace":"com.azure.schemaregistry.samples","fields":[{"name":"name","type":"string"},{"name":"favoriteNumber","type":"int"}]})
   .query(true)
-  .reply(400, {"Code":400,"Detail":"Invalid schema type for PUT request. 'not-valid' is not supported. TrackingId:126dc8df-b692-4f1f-957b-0f85b28c19ed_G6, SystemTracker:endpoint:$schemagroups/group-1/schemas/azsdk_js_test, Timestamp:2021-08-04T20:29:19"}, [
+  .reply(400, {"Code":400,"Detail":"Invalid schema type for PUT request. 'not-valid' is not supported. TrackingId:d077b0d2-33f7-4a20-a4c1-b8da6276814c_G28, SystemTracker:endpoint:$schemagroups/group-1/schemas/azsdk_js_test, Timestamp:2021-09-24T16:46:22"}, [
   'Transfer-Encoding',
   'chunked',
   'Content-Type',
@@ -123,5 +123,5 @@ nock('https://endpoint', {"encodedQueryParams":true})
   'Strict-Transport-Security',
   'max-age=31536000',
   'Date',
-  'Wed, 04 Aug 2021 20:29:19 GMT'
+  'Fri, 24 Sep 2021 16:46:22 GMT'
 ]);